feat(know-your-hazards): dynamically load and display the nearest 20 critical facilities (#164)
# Summary

When the user enters their location in the map, the nearest 20 critical facilities are displayed in the side panel. The same list is loaded in the map. When the user clicks on a critical facility in the side panel list, the map centers at the coordinates of the selected facility.

Only the critical facilities displayed in the side panel gets displayed in the map. This means the critical facilities displayed in the map varies according to the location of the user. ✨ 

# Demo

![Kapture 2021-10-01 at 04 50 42](https://user-images.githubusercontent.com/11599005/135528497-5f30a9c7-67b3-4d64-bd79-e9a6ed18aca4.gif)